added js modules for constants (const.js), mock functions (mockFunctions.js), base functions (functions.js) and utilities functions (utils.js)
