+++ index.html
@@ -10,8 +10,7 @@
 </head>
 
 <body>
-<script async src="js/functions.js"></script>
-<script async src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 <main>
 
   <!-- Фильтрация изображений от других пользователей -->
