+++ js/const.js
@@ -1,0 +1,32 @@
+export const TEST_COMMENTS = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+  ],
+  TEST_NAMES = [
+    'Кирилл',
+    'Евгений',
+    'Артем',
+    'Артур',
+    'Александр',
+    'Михаил',
+    'Ксения',
+    'Алена',
+    'Арина',
+    'Алина',
+    'Петр',
+    'Ульяна',
+    'Анна',
+    'Григорий'
+  ],
+  TEST_DESCRIPTIONS = [
+    'С волками жить - по волчьи выть.',
+    'Настоящая братва - это единицы из сотен, что своих не бросали.',
+    'Друзей выбираем мы сами, но лучших оставляет время.',
+    'Мне не важно прав он или нет, он мой брат, и я тебя за него разорву!',
+    'Своя компания - это компания, в которой можно не думать когда говоришь.',
+    'Живи, братуха, и не думай, что кто-то про тебя забыл. Друзья не забывают друга, а кто забыл, тот не был им.'
+  ];
